+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Add `matchedRoute` (e.g. `/users/:id`)  and `matchedServiceKey` (e.g. `api.users`) into the context state
+
 ## 6.4.0 (16th July 2022)
 * `@Validate` and `@Required` will now only check for `undefined` and will not count `null` properties as missing